% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrCovariance.R
\name{corrCovariance}
\alias{corrCovariance}
\title{Correlated Block Covariance Matrix}
\usage{
corrCovariance(p, inblock, outblock, numBlocks)
}
\arguments{
\item{p}{Number of dimensions.}

\item{inblock}{Covariance value inside of block.}

\item{outblock}{Covariance value outside of block.}

\item{numBlocks}{Total number of blocks. Defaults to p/25.}
}
\description{
Creates a covariance structure with a block diagonal.
}
