% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsTest2.R
\name{crsTest2}
\alias{crsTest2}
\alias{burrowClusters2}
\title{Burrow Clustered Random Subspaces}
\usage{
crsTest2(x, y, B1 = 100)

burrowClusters2(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\description{
Burrow Clustered Random Subspaces
}
