% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dempTest.R
\name{dempTest}
\alias{dempTest}
\title{Dempster Test}
\usage{
dempTest(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\value{
Dempster statistic for different mean vectors.
}
\description{
Dempster Test
}
