% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdTest.R
\name{sdTest}
\alias{sdTest}
\alias{sdPval}
\alias{sd.test}
\title{Srivastava-Du Test}
\usage{
sdTest(x, y)

sdPval(x, y)

sd.test(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\description{
Calculates Srivastava-Du test statistic for two
high-dimensional data sets.
}
