% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftedMean.R
\name{shiftedMean}
\alias{shiftedMean}
\title{Shifted Mean Vector for Data Generation}
\usage{
shiftedMean(mean, shift, m)
}
\arguments{
\item{mean}{Original mean vector.}

\item{shift}{Amount to shift by.}

\item{m}{Number of blocks to experience shift.}
}
\description{
This function shifts a given mean vector. The function shifts 4/5 of each
block in m out of the total number of blocks.
}
