% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpTest.R
\name{smpTest}
\alias{smpTest}
\alias{smpPval}
\title{Srivastava T+2 Test}
\usage{
smpTest(x, y)

smpPval(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\description{
Calculates Srivastava test statstic base on Moore-Penrose
inverse for two high dimensional data sets.
}
