% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdReducedPair2.R
\name{svdReducedPair2}
\alias{svdReducedPair2}
\title{SVD Reduction for High Dimensional Data Pairs.}
\usage{
svdReducedPair2(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\description{
The following performs dimension reduction on two samples of
high dimensional data.
}
