% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdTest.R
\name{svdTest}
\alias{svdTest}
\alias{hotellingT2}
\title{SVD Test}
\usage{
svdTest(x, y, method = "svd")

hotellingT2(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\description{
The following performs a Hotelling T2 test on a high dimensional
dataset that has first been reduced via the singular value
decomposition.
}
