% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsTest.R
\name{bsTest}
\alias{bsTest}
\title{Bai Sarandasa Test}
\usage{
bsTest(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\value{
Bai Sarandasa statistic for different mean vectors.
}
\description{
Bai Sarandasa Test
}
