% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsTest.R
\name{crsTest}
\alias{crsTest}
\alias{burrowClusters}
\title{Burrow Clustered Random Subspaces}
\usage{
crsTest(x, y, B1 = 100)

burrowClusters(x, y)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}

\item{B1}{Number of times to randomly subset. Defaults to 100.}
}
\value{
Clustered random subspaces statistic.
}
\description{
Burrow Clustered Random Subspaces
}
