% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsTest.R, R/crsTest2.R
\name{crsTest}
\alias{crsTest}
\alias{crs.test}
\alias{worleyClusters}
\title{Worley Clustered Random Subspaces}
\usage{
crsTest(x, y, B1 = 100)

crs.test(x, y, k, B1 = 100, B2 = 100)

worleyClusters(x, y)

}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}

\item{B1}{Number of times to randomly subset. Defaults to 100.}
}
\value{
Clustered random subspaces statistic.
}
\description{
Worley Clustered Random Subspaces
}
