% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csTest.R
\name{csTest}
\alias{csTest}
\alias{csPval}
\alias{hotellingCluster}
\alias{pearsonClusters}
\alias{pearsonDistance}
\alias{pearson}
\title{Zhang and Pan Cluster Subspaces Test}
\usage{
csTest(x, y)

csPval(x, y, k, B1 = 100, B2 = 100, parallel = TRUE, mc.cores = 2)

hotellingCluster(x, y)

pearsonClusters(x, y)

pearsonDistance(data)

pearson(n, p)
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}
}
\description{
Zhang and Pan Cluster Subspaces Test
}
\examples{

}
