% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdReducePair.R
\name{svdReducedPair}
\alias{svdReducedPair}
\title{SVD Reduction for High Dimensional Data Pairs.}
\usage{
svdReducedPair(x, y, method = "svd")
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}

\item{method}{Method used for Singular Value Decompostion.}
}
\description{
The following performs dimension reduction on two samples of
high dimensional data.
}
