% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdTest.R
\name{svdTest}
\alias{svdTest}
\title{SVD Test}
\usage{
svdTest(x, y, method = "svd")
}
\arguments{
\item{x}{Data set 1.}

\item{y}{Data set 2.}

\item{method}{Method for reduction. Can use traditional svd or svd of scatter matrix.}
}
\description{
The following performs a Hotelling T2 test on a high dimensional
dataset that has first been reduced via the singular value
decomposition.
}
